Document new behaviour of channel write method.
In previous versions of the module, in the channel write method, the
ugly inherited libssh2 behaviour consisting on returning EAGAIN errors
without setting the error status was used.

Now, the error status is set to LIBSSH2_ERROR_EAGAIN and the method
returns undef.

This is more perlish and consistent and greatly simplifies the code
using it at the cost of introducing a small backward incompatibility.